--- src/commandline/TopTrumps.java
@@ -1,13 +1,0 @@
-package commandline;
-
-public class TopTrumps {
-
-	public static void main(String[] args) {
-		Card card = new Card();
-		card.readStar();
-		card.shuffleCards();
-		card.ShareCards();
-
-	}
-
-}
