+++ src/commandline/TopTrumpsCLIApplication.java
@@ -1,5 +1,7 @@
 package commandline;
 
+import java.util.ArrayList;
+import java.util.Random;
 import java.util.Scanner;
 
 /**
@@ -51,25 +53,59 @@ public class TopTrumpsCLIApplication {
 		System.out.println("\nGame start\n");
 		// State
 		boolean exitGame = false; // flag to check whether the user wants to exit the game
-		Card deck = new Card();
-		deck.readStar();
-		deck.shuffleCards();
-		deck.ShareCards();
-		if(logsToFile) {
-		deck.createLogFile();
-		deck.printToFile(deck.cardLog());
-		}
+		Deck deck = new Deck(); 
+		dealCards(deck, logsToFile); 
+		// choose first player randomly
+		Random rand = new Random();
+		int firstPlay = rand.nextInt(5) + 1;
+		System.out.println(firstPlay);	
+		// initialise round 1
+		int round = 1;
+		int activePlayer = firstPlay;
 		// Loop until the user wants to exit the game
 		while (!exitGame) {
-			int rounds = 1;
-			System.out.println("Round " + rounds);
-					
-				
-					exitGame=true; // use this when the user wants to exit the game
+			
+			System.out.println("Round " + round);
+			printWhosTurn(activePlayer);
+		// code for testing		
+			System.out.println("round again?");
+			Scanner input = new Scanner(System.in);
+			String answer = input.nextLine();
+			System.out.println("who won?");
+			activePlayer = input.nextInt();
+			if(answer.equals("n")) {
+				exitGame=true; // use this when the user wants to exit the game
+				deck.printToFile(deck.cardLog());
+			}else {
+				round++;
+			}
 				}//while-end
 		
 		return;
 	}//playGame-end
+	
+	public Deck dealCards(Deck deck, boolean logsToFile) {
+	// load, shuffle, and deal the cards
+		deck.readStar();
+		deck.shuffleCards();
+		deck.ShareCards();
+		// if flag selected print info to log
+		if(logsToFile) {
+			deck.createLogFile();
+			deck.printToFile(deck.cardLog());
+		}
+		return deck;
+	}
+	
+	public void printWhosTurn(int player) {
+		if(player == 1) {
+			System.out.println("It's your turn to choose a category");
+		}else {
+			System.out.println("It's the computers turn to choose a category");
+		}
+	}
+	
+	
 	
 	public void printStats() {
 //--------------------------------
